Fix some issues in 'storage.py' file.
Line 28: 'template_name' and 'template_file' are two compulsory parameters that were not
passed in the fromfile() method.

Line 69: There is no test() method defined in the deployment module
(monster/deployments/deployment.py). We will fix this in a separate pull request

Line 96: The from_chef_environment() method only expects one parameter which is the environmemt.